requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
cmasher==1.6.0
cmocean==2.0
colorama==0.4.4
colorspacious==1.1.2
-e git+https://github.com/JosiePark/TTI_impact.git@6af1b154a1eb670238e6ff6667b18cbcb7beeb9a#egg=covidTTI
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
dill==0.3.3
e13tools==0.9.6
entrypoints==0.3
et-xmlfile==1.1.0
gitdb==4.0.7
GitPython==3.1.14
graphviz==0.16
ipykernel==5.5.4
ipython==7.23.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
jdcal==1.4.1
jedi==0.18.0
jellyfish==0.8.2
Jinja2==2.11.3
jsonpickle==2.0.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
llvmlite==0.31.0
MarkupSafe==1.1.1
matplotlib==3.4.1
matplotlib-inline==0.1.2
mistune==0.8.4
multiprocess==0.70.11.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.3.0
numba==0.48.0
numpy==1.20.2
openpyexcel==2.5.14
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
psutil==5.8.0
pycparser==2.20
pydot==1.4.2
Pygments==2.9.0
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pywinpty==1.0.1
PyYAML==5.4.1
pyzmq==22.0.3
qtconsole==5.1.0
QtPy==1.9.0
scipy==1.6.3
sciris==1.1.1
seaborn==0.11.1
Send2Trash==1.5.0
six==1.16.0
smmap==4.0.0
synthpops==1.0.0
terminado==0.9.4
testpath==0.4.4
tornado==6.1
traitlets==5.0.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
wincertstore==0.2
xlrd==1.2.0
XlsxWriter==1.4.2
